Always perform clean up on application on DAS upon undeployment.
EarLibClassloader of an EAR app deployed on non-DAS instance was initialized and not closed on DAS upon deployment.
That lead to JarFile handles to its lib directory being held, and the DAS was unable to clear its application directory.
Calling unload unconditionally gurantees ApplicationInfo.clean to be called, which in turns calls done() on all application
class loaders.